Added exception handling

+++ src/metadata/service/developer.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from "@nestjs/common";
+import { BadRequestException, Injectable, NotFoundException } from "@nestjs/common";
 import { InjectRepository } from '@nestjs/typeorm';
 import { PageOptions } from "src/common/page.dto";
 import { DeepPartial, Repository } from 'typeorm';
@@ -12,39 +12,75 @@ export class DeveloperService
     private developerRepository: Repository<Developer>;
 
 
-    public create(developer: Developer): Promise<Developer>
+    public async create(developer: Developer): Promise<Developer>
     {
-        return this.developerRepository.save(developer);
+        try {
+            return await this.developerRepository.save(developer);
+
+        } catch(exception) {
+
+            throw new BadRequestException("Body: data is invalid");
+        }
     }
 
     public retrieve(id: number): Promise<Developer>
     {
-        return this.developerRepository.findOneBy({
-            id: id
-        });
+        const entity = this.developerRepository.findOneBy({ id: id });
+
+        if(!entity) {
+            throw new NotFoundException(`Entity with ID ${id} does not exist`);
+        }
+
+        return entity;
     }
 
     public retrieveByUserName(userName: string): Promise<Developer>
     {
-        return this.developerRepository.findOneBy({
-            userName: userName
-        });
+        const entity = this.developerRepository.findOneBy({ userName: userName });
+
+        if(!entity) {
+            throw new NotFoundException(`Entity with user name ${userName} does not exist`);
+        }
+
+        return entity;
     }
 
     public retrieveByEmail(email: string): Promise<Developer>
     {
-        return this.developerRepository.findOneBy({
-            email: email
-        });
+        const entity = this.developerRepository.findOneBy({ email: email });
+
+        if(!entity) {
+            throw new NotFoundException(`Entity with email ${email} does not exist`);
+        }
+
+        return entity;
     }
 
-    public update(id: number, developer: DeepPartial<Developer>)
+    public async update(id: number, developer: DeepPartial<Developer>)
     {
-        this.developerRepository.update({ id: id }, developer);
+        const exists = this.developerRepository.findOneBy({ id: id });
+
+        if(!exists) {
+            throw new NotFoundException(`Entity with ID ${id} does not exist`);
+        }
+
+        try {
+            await this.developerRepository.update({ id: id }, developer);
+
+        } catch(exception) {
+
+            throw new BadRequestException("Body: data is invalid");
+        }
     }
 
     public delete(id: number)
     {
+        const exists = this.developerRepository.findOneBy({ id: id });
+
+        if(!exists) {
+            throw new NotFoundException(`Entity with ID ${id} does not exist`);
+        }
+
         this.developerRepository.delete({ id: id });
     }
